Bug 913228 - Fix floating point math to not round to zero accidentally. r=autra
UltraBlame original commit: b5eae5bcf096869baf8f35384b6ba36d05b4ca69

+++ mobile/android/base/gfx/Axis.java
@@ -92,7 +92,7 @@ abstract class Axis {
 
     
     static float getFrameAdjustedFriction(float baseFriction, long currentNsPerFrame) {
-        float framerateMultiplier = currentNsPerFrame / NS_PER_FRAME;
+        float framerateMultiplier = (float)currentNsPerFrame / NS_PER_FRAME;
         return (float)Math.pow(Math.E, (Math.log(baseFriction) / framerateMultiplier));
     }
